Rename folder and change component structure
