+++ src/events/components/DetailView/AttendanceEvent.tsx
@@ -15,14 +15,15 @@ import Attendance from '../Attendance';
 import { attendeeSelectors, fetchAttendeeByEventId } from 'events/slices/attendees';
 import EventPaymentBlock from '../EventPayment/EventPaymentBlock';
 import EventPrice from '../EventPayment/EventPrice';
-import { PublicAttendees } from './PublicAttendees';
+import { PublicAttendees } from './PublicAttendeesModal/PublicAttendeesWrapper';
+import ParticipantsButton from './PublicAttendeesModal/ParticipantsButton';
 
 interface IProps {
   eventId: number;
   eventTitle: string;
 }
 
-const AttendanceEvent: FC<IProps> = ({ eventId, eventTitle}) => {
+const AttendanceEvent: FC<IProps> = ({ eventId, eventTitle }) => {
   const dispatch = useDispatch();
   const attendanceEvent = useSelector((state) => attendanceEventSelectors.selectById(state, eventId));
   const attendee = useSelector((state) => attendeeSelectors.selectAll(state)).find(
@@ -71,7 +72,9 @@ const AttendanceEvent: FC<IProps> = ({ eventId, eventTitle}) => {
       </Block>
       <div className={`${style.attendanceContainer} ${style.fullBlock}`}>
         <Attendance canAttend={isEligibleForSignup} event={attendanceEvent} unattendDeadline={cancellationDeadline} />
-        <PublicAttendees eventTitle={eventTitle} eventId={eventId} isAttending={attendanceEvent.is_attendee} canAttend={isEligibleForSignup} />
+        <PublicAttendees isAttending={attendanceEvent.is_attendee} canAttend={isEligibleForSignup}>
+          <ParticipantsButton eventTitle={eventTitle} eventId={eventId} />
+        </PublicAttendees>
       </div>
       {attendanceEvent.payment && (
         <Block title="Pris" className={`${style.fullBlock} ${style.priceBlock}`}>
