+++ src/events/components/DetailView/PublicAttendeesModal/Attendee.less
