+++ src/events/components/DetailView/PublicAttendeesModal/Attendee.tsx
