+++ src/events/components/DetailView/PublicAttendeesModal/ParticipantsButton.tsx
