+++ src/events/components/DetailView/PublicAttendeesModal/PublicAttendees.less
