+++ src/events/components/DetailView/PublicAttendeesModal/PublicAttendeesWrapper.tsx
@@ -1,15 +1,14 @@
+import React, { FC, ReactNode } from 'react';
 import { ISignupEligibility } from 'events/models/Event';
-import React, { FC } from 'react';
-import ParticipantsButton from './ParticipantsButton';
+
 interface IProps {
-  eventId: number;
-  eventTitle: string;
   isAttending: boolean;
   canAttend: ISignupEligibility | null;
+  children?: ReactNode;
 }
 
-export const PublicAttendees: FC<IProps> = ({ eventId, eventTitle, isAttending, canAttend }) => {
+export const PublicAttendees: FC<IProps> = ({ children, isAttending, canAttend }) => {
   if (!canAttend) return null; // The user is not logged in
   if (!isAttending && !canAttend.status) return null; // The user is not attending, or the user is not allowed to attend the event
-  return <ParticipantsButton eventId={eventId} eventTitle={eventTitle} />;
+  return <>{children}</>;
 };
